// src/app/modules/academicSemester/academicSemester.interface.ts
export type IAcademicSemesterFilters = {
  searchTerm?: string;
};

export type IAcamedicSemesterTitles = 'Autumn' | 'Summer' | 'Fall';
export type IAcamedicSemesterCodes = '01' | '02' | '03' | '04';
export type IAcademicSemesterMonths =
  | 'January'
  | 'February'
  | 'March'
  | 'April'
  | 'May'
  | 'June'
  | 'July'
  | 'August'
  | 'September'
  | 'October'
  | 'November'
  | 'December';